move perceval/algorithm to perceval/utils/algorithms
