+++ perceval/utils/algorithms/decomposition.py
