+++ perceval/utils/algorithms/norm.py
