+++ perceval/utils/algorithms/solve.py
